requirements.txt
Flask==3.0.3
google-api-python-client==2.130.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
pdfplumber==0.11.0
PyPDF2==3.0.1
python-dotenv==1.0.1